// src/bcmr/bcmr-v2.schema.ts
export type URIs = {
  [identifier: string]: string;
};
// ...
export type Extensions = {
  [extensionIdentifier: string]:
    | string
    | { [key: string]: string }
    | { [keyA: string]: { [keyB: string]: string } };
};
// ...
export type Tag = {
  /**
   * The name of this tag for use in interfaces.
   *
   * In user interfaces with limited space, names should be hidden beyond
   * the first newline character or `20` characters until revealed by the user.
   *
   * E.g.:
   * - `Individual`
   * - `Token`
   * - `Audited by ACME, Inc.`
   */
  name: string;

  /**
   * A string describing this tag for use in user interfaces.
   *
   * In user interfaces with limited space, descriptions should be hidden beyond
   * the first newline character or `140` characters until revealed by the user.
   *
   * E.g.:
   * - `An identity maintained by a single individual.`
   * - `An identity representing a type of token.`
   * - `An on-chain application that has passed security audits by ACME, Inc.`
   */
  description?: string;

  /**
   * A mapping of identifiers to URIs associated with this tag. URI identifiers
   * may be widely-standardized or registry-specific. Values must be valid URIs,
   * including a protocol prefix (e.g. `https://` or `ipfs://`). Clients are
   * only required to support `https` and `ipfs` URIs, but any scheme may
   * be specified.
   *
   * The following identifiers are recommended for all tags:
   * - `icon`
   * - `web`
   *
   * The following optional identifiers are standardized:
   * - `blog`
   * - `chat`
   * - `forum`
   * - `icon-intro`
   * - `registry`
   * - `support`
   *
   * For details on these standard identifiers, see:
   * https://github.com/bitjson/chip-bcmr#uri-identifiers
   *
   * Custom URI identifiers allow for sharing social networking profiles, p2p
   * connection information, and other application-specific URIs. Identifiers
   * must be lowercase, alphanumeric strings, with no whitespace or special
   * characters other than dashes (as a regular expression: `/^[-a-z0-9]+$/`).
   *
   * For example, some common identifiers include: `discord`, `docker`,
   * `facebook`, `git`, `github`, `gitter`, `instagram`, `linkedin`, `matrix`,
   * `npm`, `reddit`, `slack`, `substack`, `telegram`, `twitter`, `wechat`,
   * `youtube`.
   */
  uris?: URIs;

  /**
   * A mapping of `Tag` extension identifiers to extension definitions.
   * {@link Extensions} may be widely standardized or application-specific.
   */
  extensions?: Extensions;
};
// ...
export type SequentialNftCollection = {
  /**
   * A mapping of each NFT commitment (typically, a positive integer encoded as
   * a VM number) to metadata for that NFT type in this category.
   */
  types: {
    /**
     * Interpretation information for each type of NFT within the token
     * category, indexed by commitment hex. For sequential NFTs, the on-chain
     * commitment of each NFT is interpreted as a VM number to reference its
     * particular NFT type in user interfaces. Issuing a sequential NFT with a
     * negative or invalid VM number is discouraged, but clients may render the
     * commitment of such NFTs in hex-encoded form, prefixed with `X`.
     */
    [commitmentHex: string]: NftType;
  };
};
// ...
export type NftCategory = {
  /**
   * A string describing how this identity uses NFTs (for use in user
   * interfaces). Descriptions longer than `160` characters may be elided in
   * some interfaces.
   *
   * E.g.:
   * - "ACME DEX NFT order receipts are issued when you place orders on the
   * decentralized exchange. After orders are processed, order receipts can
   * be redeemed for purchased tokens or sales proceeds.";
   * - "ACME Game collectable NFTs unlock unique playable content, user
   * avatars, and item skins in ACME Game Online."; etc.
   */
  description?: string;

  /**
   * A mapping of field identifier to field definitions for the data fields
   * that can appear in NFT commitments of this category.
   *
   * Categories including only sequential NFTs (where `parse.bytecode` is
   * undefined) should omit `fields` (or set to `undefined`).
   */
  fields?: NftCategoryField;

  /**
   * Parsing and interpretation information for all NFTs of this category;
   * this enables generalized wallets to parse and display detailed
   * information about all NFTs held by the wallet, e.g. `BCH Pledged`,
   * `Order Price`, `Seat Number`, `Asset Number`,
   * `IPFS Content Identifier`, `HTTPS URL`, etc.
   *
   * Parsing instructions are provided in the `bytecode` property, and the
   * results are interpreted using the `types` property.
   */
  parse: SequentialNftCollection | ParsableNftCollection;
};

/**
 * A definition specifying information about an identity's token category.
 */
export type TokenCategory = {
  /**
   * The current token category used by this identity. Often, this will be
   * equal to the identity's authbase, but some token identities must migrate
   * to new categories for technical reasons.
   */
  category: string;

  /**
   * An abbreviation used to uniquely identity this token category.
   *
   * Symbols must be comprised only of capital letters, numbers, and dashes
   * (`-`). This can be validated with the regular expression:
   * `/^[-A-Z0-9]+$/`.
   */
  symbol: string;

  /**
   * An integer between `0` and `18` (inclusive) indicating the divisibility
   * of the primary unit of this token category.
   *
   * This is the number of digits that can appear after the decimal separator
   * in fungible token amounts. For a token category with a `symbol` of
   * `SYMBOL` and a `decimals` of `2`, a fungible token amount of `12345`
   * should be displayed as `123.45 SYMBOL`.
   *
   * If omitted, defaults to `0`.
   */
  decimals?: number;

  /**
   * Display information for non-fungible tokens (NFTs) of this identity.
   * Omitted for token categories without NFTs.
   */
  nfts?: NftCategory;
};

/**
 * A snapshot of the metadata for a particular identity at a specific time.
 */
export type IdentitySnapshot = {
  /**
   * The name of this identity for use in interfaces.
   *
   * In user interfaces with limited space, names should be hidden beyond
   * the first newline character or `20` characters until revealed by the user.
   *
   * E.g. `ACME Class A Shares`, `ACME Registry`, `Satoshi Nakamoto`, etc.
   */
  name: string;

  /**
   * A string describing this identity for use in user interfaces.
   *
   * In user interfaces with limited space, descriptions should be hidden beyond
   * the first newline character or `140` characters until revealed by the user.
   *
   * E.g.:
   * - `The common stock issued by ACME, Inc.`
   * - `A metadata registry maintained by Company Name, the embedded registry for Wallet Name.`
   * - `Software developer and lead maintainer of Wallet Name.`
   */
  description?: string;

  /**
   * An array of `Tag` identifiers marking the `Tag`s associated with this
   * identity. All specified tag identifiers must be defined in the registry's
   * `tags` mapping.
   */
  tags?: string[];

  /**
   * The timestamp at which this identity snapshot is fully in effect. This
   * value should only be provided if the snapshot takes effect over a period
   * of time (e.g. an in-circulation token identity is gradually migrating to
   * a new category). In these cases, clients should gradually migrate to
   * using the new information beginning after the identity snapshot's timestamp
   * and the `migrated` time.
   *
   * This timestamp must be provided in simplified extended ISO 8601 format, a
   * 24-character string of format `YYYY-MM-DDTHH:mm:ss.sssZ` where timezone is
   * zero UTC (denoted by `Z`). Note, this is the format returned by ECMAScript
   * `Date.toISOString()`.
   */
  migrated?: string;

  /**
   * If this identity is a type of token, a data structure indicating how tokens
   * should be understood and displayed in user interfaces. Omitted for
   * non-token identities.
   */
  token?: TokenCategory;

  /**
   * The status of this identity, must be `active`, `inactive`, or `burned`. If
   * omitted, defaults to `active`.
   * - Identities with an `active` status should be actively tracked by clients.
   * - Identities with an `inactive` status may be considered for archival by
   * clients and may be removed in future registry versions.
   * - Identities with a `burned` status have been destroyed by setting the
   * latest identity output to a data-carrier output (`OP_RETURN`), permanently
   * terminating the authchain. Clients should archive burned identities and –
   * if the burned identity represented a token type – consider burning any
   * remaining tokens of that category to reclaim funds from those outputs.
   */
  status?: 'active' | 'burned' | 'inactive';

  /**
   * The split ID of this identity's chain of record.
   *
   * If undefined, defaults to {@link Registry.defaultChain}.
   */
  splitId?: string;

  /**
   * A mapping of identifiers to URIs associated with this identity. URI
   * identifiers may be widely-standardized or registry-specific. Values must be
   * valid URIs, including a protocol prefix (e.g. `https://` or `ipfs://`).
   * Clients are only required to support `https` and `ipfs` URIs, but any
   * scheme may be specified.
   *
   * The following identifiers are recommended for all identities:
   * - `icon`
   * - `web`
   *
   * The following optional identifiers are standardized:
   * - `blog`
   * - `chat`
   * - `forum`
   * - `icon-intro`
   * - `image`
   * - `migrate`
   * - `registry`
   * - `support`
   *
   * For details on these standard identifiers, see:
   * https://github.com/bitjson/chip-bcmr#uri-identifiers
   *
   * Custom URI identifiers allow for sharing social networking profiles, p2p
   * connection information, and other application-specific URIs. Identifiers
   * must be lowercase, alphanumeric strings, with no whitespace or special
   * characters other than dashes (as a regular expression: `/^[-a-z0-9]+$/`).
   *
   * For example, some common identifiers include: `discord`, `docker`,
   * `facebook`, `git`, `github`, `gitter`, `instagram`, `linkedin`, `matrix`,
   * `npm`, `reddit`, `slack`, `substack`, `telegram`, `twitter`, `wechat`,
   * `youtube`.
   */
  uris?: URIs;

  /**
   * A mapping of `IdentitySnapshot` extension identifiers to extension
   * definitions. {@link Extensions} may be widely standardized or
   * application-specific.
   *
   * Standardized extensions for `IdentitySnapshot`s include the `authchain`
   * extension. See
   * https://github.com/bitjson/chip-bcmr#authchain-extension for details.
   */
  extensions?: Extensions;
};

/**
 * A snapshot of the metadata for a particular chain/network at a specific
 * time. This allows for registries to provide similar metadata for each chain's
 * native currency unit (name, description, symbol, icon, etc.) as can be
 * provided for other registered tokens.
 */
export type ChainSnapshot = Omit<IdentitySnapshot, 'migrated' | 'token'> & {
  /**
   * A data structure indicating how the chain's native currency units should be
   * displayed in user interfaces.
   */
  token: {
    /**
     * An abbreviation used to uniquely identity this native currency unit.
     *
     * Symbols must be comprised only of capital letters, numbers, and dashes
     * (`-`). This can be validated with the regular expression:
     * `/^[-A-Z0-9]+$/`.
     */
    symbol: string;

    /**
     * An integer between `0` and `18` (inclusive) indicating the divisibility
     * of the primary unit of this native currency.
     *
     * This is the number of digits that can appear after the decimal separator
     * in currency amounts. For a currency with a `symbol` of `SYMBOL` and a
     * `decimals` of `2`, an amount of `12345` should be displayed as
     * `123.45 SYMBOL`.
     *
     * If omitted, defaults to `0`.
     */
    decimals?: number;
  };
};
// ...
export type RegistryTimestampKeyedValues<T> = {
  [timestamp: string]: T;
};

/**
 * A block height-keyed map of {@link ChainSnapshot}s documenting the evolution
 * of a particular chain/network's identity. Like {@link IdentityHistory}, this
 * structure allows wallets and other user interfaces to offer better
 * experiences when a chain identity is rebranded, redenominated, or other
 * important metadata is modified in a coordinated update.
 */
export type ChainHistory = RegistryTimestampKeyedValues<ChainSnapshot>;
// ...
export type IdentityHistory = RegistryTimestampKeyedValues<IdentitySnapshot>;

/**
 * An identity representing a metadata registry that is not published on-chain
 * and therefore has no authbase or trackable authchain.
 */
export type OffChainRegistryIdentity = Pick<
  IdentitySnapshot,
  'name' | 'description' | 'uris' | 'tags' | 'extensions'
>;